requirements.txt
aiogram==3.6.0
pydantic-settings==2.2.1
Pillow==10.4.0
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.30
greenlet==3.1.1
aiohttp==3.9.5
python-dotenv==1.0.1
